+++ velithon/responses/plain_text.py
@@ -1,0 +1,9 @@
+"""Plain text Response implementation."""
+
+from .base import Response
+
+
+class PlainTextResponse(Response):
+    """Response for plain text content."""
+    
+    media_type = 'text/plain'
